services test done

+++ services/product management/api.py
@@ -1,4 +1,4 @@
-from fastapi import FastAPI
+from fastapi import FastAPI, HTTPException
 from DataBaseManagement import DataBase
 import uvicorn
 from Models import *
@@ -14,14 +14,26 @@ async def get_items():
 
 @app.post("/")
 async def add_product(item: ProductModel):
-    return d.add_product(item)
+    if d.is_product_exist(item):
+        raise HTTPException(status_code=400, detail="Product already exists")
+    inserted_id = d.add_product(item)
+    item.id = inserted_id
+    return item.dict()
 
 
 @app.put("/")
 async def update_product(item: ProductModel):
-    return d.update_product(item)
+    if d.is_product_exist(item):
+        raise HTTPException(status_code=400, detail="Product already exists")
+    if d.get_product_by_id(item.id) is None:
+        raise HTTPException(status_code=404, detail="product not found")
+    d.update_product(item)
+    return item.dict()
 
 
 @app.delete("/")
 async def delete_product(item_id: int):
-    return d.delete_product(item_id)
+    if d.get_product_by_id(item_id) is None:
+        raise HTTPException(status_code=404, detail="product not found")
+    d.delete_product(item_id)
+    return {"message": "Product deleted"}
